Small fixes to allow unit tests to compile with gcc 4.7.x (tested with gcc 4.7.2)
Review URL: https://chromiumcodereview.appspot.com/11568033

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@174956 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 1885369433d7d51a164afbe73758752eb6c753d6

+++ base/string_number_conversions_unittest.cc
@@ -203,7 +203,7 @@ TEST(StringNumberConversionsTest, StringToInt64) {
 TEST(StringNumberConversionsTest, HexStringToInt) {
   static const struct {
     std::string input;
-    int output;
+    int64 output;
     bool success;
   } cases[] = {
     {"0", 0, true},
